+++ spec/dummy/spec/helpers/react_on_rails_pro_helper_spec.rb
@@ -9,7 +9,7 @@ describe ReactOnRailsProHelper, type: :helper do
   # In order to test the pro helper, we need to load the methods from the regular helper.
   # I couldn't see any easier way to do this.
   include ReactOnRails::Helper
-  include Webpacker::Helper
+  include Shakapacker::Helper
   before do
     allow(self).to receive(:request) {
       RequestDetails.new("http://foobar.com/development", { "HTTP_ACCEPT_LANGUAGE" => "en" })